+++ src/components/UploadCard/processFile.ts
@@ -1,4 +1,4 @@
-import { readInChunks, ValidationAction } from '@context'
+import { ValidationAction } from '@context'
 import { Dispatch } from 'react'
 
 interface processFileProps {
@@ -8,5 +8,13 @@ interface processFileProps {
 }
 
 export const processFile = (props: processFileProps) => {
-  readInChunks(props)
+  const { fileId, dispatch, file } = props
+
+  dispatch({ type: 'SET_FILE', payload: fileId, fileId })
+  const worker = new Worker('/src/context/ValidationContext/worker.ts', { type: 'module' })
+  worker.postMessage({ file, fileId })
+  worker.onmessage = (e) => {
+    //console.log("Message received from worker", e.data);
+    dispatch(e.data)
+  }
 }
