+++ tsconfig.json
@@ -2,7 +2,12 @@
   "compilerOptions": {
     "target": "ES2020",
     "useDefineForClassFields": true,
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
+    "lib": [
+      "ES2020",
+      "DOM",
+      "DOM.Iterable",
+      "webworker"
+    ],
     "module": "ESNext",
     "skipLibCheck": true,
 
@@ -28,7 +33,7 @@
       "@context": ["context"],
       "@pages": ["pages"],
       "@mocks": ["__mocks__"],
-      "@test": ["__test__"],
+      "@test": ["__test__"]
     }
   },
   "include": ["src"],
